feat(parse): add ast
