+++ crates/claw_parse/src/error.rs
@@ -1,0 +1,3 @@
+use chumsky::error::Simple;
+
+pub type ParseError = Simple<char>;
