+++ crates/claw_parse/src/parser/mod.rs
@@ -1,0 +1,7 @@
+use chumsky::prelude::*;
+use crate::{AST, ParseError};
+
+pub fn create_parser() -> impl Parser<char, AST, Error = ParseError> {
+    empty()
+        .then_ignore(end())
+}
